move scene conversion registration patch and check applicable types properly
